minor improvements and bug fixes

+++ cmds/whitelist.js
@@ -16,7 +16,7 @@ module.exports.run = async (Client, message, args) => {
 		let msg = await message.channel.send(`Adding ${usr} to the whitelist...`);
 		/* The most scuffed use of exec in my entire life */
 		exec(
-			`tempfile=$(mktemp) && echo 'whitelist add ${usr}' | cat > $tempfile && screen -X readbuf $tempfile && screen -X paste . && rm -f $tempfile`,
+			`tempfile=$(mktemp) && echo 'whitelist add ${usr}' | cat > $tempfile && screen -X readbuf $tempfile && screen -X paste . && rm -f $tempfile && sleep 1 && tail -1 /home/bryson/minecraftserver/logs/latest.log | cut -d ":" -f 4`,
 			(error, stdout, stderr) => {
 				if (error) {
 					return message.channel.send(`Failed! Here's what we know: ${error}`);
@@ -24,14 +24,18 @@ module.exports.run = async (Client, message, args) => {
 				if (stderr) {
 					return message.channel.send(`stderr! ${stderr}`);
 				}
+				if (stdout) {
+					const embed = new Discord.RichEmbed()
+						.setColor(0x5d2079)
+						.setTitle('Notice:')
+						.setDescription(`${stdout}`);
+					message.channel
+						.send(embed)
+						.catch((error) => message.reply(`${error}`));
+					msg.delete();
+				}
 			}
 		);
-		const embed = new Discord.RichEmbed()
-			.setColor(0x5d2079)
-			.setTitle('Success!')
-			.setDescription(`${usr} was whitelisted on the minecraft server!`);
-		message.channel.send(embed).catch((error) => message.reply(`${error}`));
-		msg.delete();
 	} else if (type === 'remove') {
 		if (!usr) {
 			return message.reply('please specify a Minecraft username!');
@@ -41,14 +45,11 @@ module.exports.run = async (Client, message, args) => {
 				'only an admin of this discord can remove people from whitelist!'
 			);
 		}
-		if (!args[1]) {
-			return message.reply('please specify a Minecraft username!');
-		}
 		let msg = await message.channel.send(
 			`Removing ${usr} from the whitelist...`
 		);
 		exec(
-			`tempfile=$(mktemp) && echo 'whitelist remove ${usr}' | cat > $tempfile && screen -X readbuf $tempfile && screen -X paste . && rm -f $tempfile`,
+			`tempfile=$(mktemp) && echo 'whitelist remove ${usr}' | cat > $tempfile && screen -X readbuf $tempfile && screen -X paste . && rm -f $tempfile && sleep 1 && tail -1 /home/bryson/minecraftserver/logs/latest.log | cut -d ":" -f 4`,
 			(error, stdout, stderr) => {
 				if (error) {
 					return message.channel.send(`Failed! Here's what we know: ${error}`);
@@ -56,23 +57,26 @@ module.exports.run = async (Client, message, args) => {
 				if (stderr) {
 					return message.channel.send(`stderr! ${stderr}`);
 				}
+				if (stdout) {
+					const embed = new Discord.RichEmbed()
+						.setColor(0x5d2079)
+						.setTitle('Notice:')
+						.setDescription(`${stdout}`);
+					message.channel
+						.send(embed)
+						.catch((error) => message.reply(`${error}`));
+					msg.delete();
+				}
 			}
 		);
-		const embed = new Discord.RichEmbed()
-			.setColor(0x5d2079)
-			.setTitle('Success!')
-			.setDescription(
-				`${usr} was removed from the whitelist on the minecraft server!`
-			);
-		message.channel.send(embed).catch((error) => message.reply(`${error}`));
-		msg.delete();
-	} else if (type === 'list') {
-    /* 
+	} else if (type === undefined) {
+		/* 
     Change the directory of your minecraft server.
     Am very proud of this tho. Works like a charm :^)
     */
+		let msg = await message.channel.send(`Getting whitelist...`);
 		exec(
-			`tempfile=$(mktemp) && echo 'whitelist list' | cat > $tempfile && screen -X readbuf $tempfile && screen -X paste . && rm -f $tempfile && sleep 2 && tail -1 /home/bryson/minecraftserver/logs/latest.log | cut -d ":" -f 5 | tr , '\n' | sort`,
+			`tempfile=$(mktemp) && echo 'whitelist list' | cat > $tempfile && screen -X readbuf $tempfile && screen -X paste . && rm -f $tempfile && sleep 1 && tail -1 /home/bryson/minecraftserver/logs/latest.log | cut -d ":" -f 5 | tr , '\n' | sort`,
 			(error, stdout, stderr) => {
 				if (error) {
 					return message.channel.send(`Failed! Here's what we know: ${error}`);
@@ -84,17 +88,20 @@ module.exports.run = async (Client, message, args) => {
 					const embed = new Discord.RichEmbed()
 						.setColor(0x5d2079)
 						.setTitle('Whitelist:')
-            .setDescription(`${stdout}`);
-            message.channel.send(embed).catch((error) => message.reply(`${error}`));
+						.setDescription(`${stdout}`);
+					message.channel
+						.send(embed)
+						.catch((error) => message.reply(`${error}`));
+					msg.delete();
 				}
 			}
 		);
 	} else {
-		return message.reply('Failed! Talk to the guy running the bot!');
+		return message.reply('Invalid input! Check your command usage!');
 	}
 };
 
 module.exports.help = {
 	name: 'whitelist',
-	usage: 'whitelist add/remove/list <minecraft username>',
+	usage: 'whitelist [add/remove] [minecraft username]',
 };
